requirements.txt
ollama
streamlit
python-dotenv
PyPDF2
langchain-text-splitters
google-api-python-client==2.161.0
google-genai